+++ Cargo.toml
@@ -1,16 +1,12 @@
 [package]
 name = "pubkycore"
-version = "0.1.0"
+version = "0.1.1"
 edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [lib]
 crate_type = ["cdylib"]
 name = "pubkycore"
-
-[[bin]]
-name = "testing"
-path = "testing/main.rs"
 
 [net]
 git-fetch-with-cli = true
@@ -25,9 +21,9 @@ tokio = "1.40.0"
 url = "2.5.2"
 base64 = "0.22.1"
 once_cell = "1.19.0"
-pubky = "0.3.0"
-pkarr = "2.2.1-alpha.2"
-pubky-common = "0.1.0"
+pubky = "0.4.0"
+pkarr = "3.3.1"
+pubky-common = "0.3.0"
 
 [dev-dependencies]
 tokio = { version = "1.40.0", features = ["full"] }
